requirements.txt
asn1crypto
cachetools
requests
python-dotenv
cryptography
elasticsearch>=8.0.0